requirements.txt
streamlit==1.30.0
pandas==2.1.3
pillow==10.1.0
requests==2.31.0
